+++ plugins/scaffolder-backend/src/index.ts
@@ -24,6 +24,8 @@ export * from './scaffolder';
 export * from './service/router';
 export * from './lib';
 export * from './processor';
-export * from './extension';
+export * from './modules';
 export { scaffolderPlugin } from './ScaffolderPlugin';
 export type { ScaffolderPluginOptions } from './ScaffolderPlugin';
+
+export * from './deprecated';