+++ plugins/adr-common/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @backstage/plugin-adr-common
 
+## 0.2.6-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/catalog-model@1.1.6-next.0
+
 ## 0.2.5
 
 ### Patch Changes