+++ plugins/allure/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage/plugin-allure
 
+## 0.1.31-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog-react@1.3.0-next.0
+  - @backstage/catalog-model@1.1.6-next.0
+
 ## 0.1.30
 
 ### Patch Changes