+++ plugins/bazaar-backend/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-bazaar-backend
 
+## 0.2.5-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-test-utils@0.1.34-next.0
+  - @backstage/backend-common@0.18.2-next.0
+  - @backstage/plugin-auth-node@0.2.11-next.0
+
 ## 0.2.3
 
 ### Patch Changes