+++ plugins/catalog-backend-module-github/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @backstage/plugin-catalog-backend-module-github
 
+## 0.2.5-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/catalog-model@1.1.6-next.0
+  - @backstage/backend-common@0.18.2-next.0
+  - @backstage/catalog-client@1.3.1-next.0
+  - @backstage/plugin-catalog-backend@1.7.2-next.0
+  - @backstage/plugin-catalog-common@1.0.11-next.0
+  - @backstage/plugin-catalog-node@1.3.3-next.0
+  - @backstage/backend-tasks@0.4.3-next.0
+  - @backstage/backend-plugin-api@0.3.2-next.0
+  - @backstage/plugin-events-node@0.2.3-next.0
+
 ## 0.2.3
 
 ### Patch Changes