+++ plugins/catalog-backend-module-ldap/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-catalog-backend-module-ldap
 
+## 0.5.9-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/catalog-model@1.1.6-next.0
+  - @backstage/plugin-catalog-backend@1.7.2-next.0
+  - @backstage/backend-tasks@0.4.3-next.0
+
 ## 0.5.7
 
 ### Patch Changes