+++ plugins/catalog-import/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-catalog-import
 
+## 0.9.5-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog-react@1.3.0-next.0
+  - @backstage/catalog-model@1.1.6-next.0
+  - @backstage/catalog-client@1.3.1-next.0
+  - @backstage/plugin-catalog-common@1.0.11-next.0
+  - @backstage/integration-react@1.1.9
+
 ## 0.9.4
 
 ### Patch Changes