+++ plugins/fossa/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage/plugin-fossa
 
+## 0.2.47-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog-react@1.3.0-next.0
+  - @backstage/catalog-model@1.1.6-next.0
+
 ## 0.2.46
 
 ### Patch Changes