+++ plugins/github-issues/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-github-issues
 
+## 0.2.4-next.0
+
+### Patch Changes
+
+- 3f88ae9d0a: Updated README.md examples to use correct components and fixed some syntax errors.
+- Updated dependencies
+  - @backstage/plugin-catalog-react@1.3.0-next.0
+  - @backstage/catalog-model@1.1.6-next.0
+
 ## 0.2.3
 
 ### Patch Changes