+++ plugins/search-backend-module-elasticsearch/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @backstage/plugin-search-backend-module-elasticsearch
 
+## 1.1.3-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-search-backend-node@1.1.3-next.0
+
 ## 1.1.1
 
 ### Patch Changes