+++ plugins/user-settings-backend/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-user-settings-backend
 
+## 0.1.6-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/catalog-model@1.1.6-next.0
+  - @backstage/backend-common@0.18.2-next.0
+  - @backstage/plugin-auth-node@0.2.11-next.0
+
 ## 0.1.4
 
 ### Patch Changes