+++ plugins/vault-backend/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-vault-backend
 
+## 0.2.8-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-test-utils@0.1.34-next.0
+  - @backstage/backend-common@0.18.2-next.0
+  - @backstage/backend-tasks@0.4.3-next.0
+
 ## 0.2.6
 
 ### Patch Changes